+++ src/components/slide-show/slide-show-interface.ts
@@ -7,4 +7,6 @@ export interface Slide {
 
 export interface SlideshowSettings {
   showThumbs?: boolean
+  ssizeLandscape?: {width: string, height: string}
+  ssizePortrait?: {width: string, height: string}
 }